# design/openapi.yaml
openapi: 3.1.0
info:
  title: Infinitude Modern Proxy API
  version: 1.0.0-draft
  summary: Northbound HTTP API for the modernized Infinitude proxy.
  description: |
    First-class, OpenAPI-described HTTP API served by the Infinitude add-on
    for consumption by the Home Assistant integration and the web UI.

    Design goals:
      * Typed resources (no array-wrapped scalars, no stringly-typed booleans).
      * Resource-oriented paths with consistent verbs.
      * Server-Sent Events push channel for state changes.
      * Comprehensive health probe.

    Out of scope for this API (handled elsewhere in the add-on):
      * The southbound protocol the thermostat uses to POST telemetry and pull changes.
      * The Carrier cloud passthrough on the `pass_reqs` cadence.

    See DESIGN.md for architecture and migration plan.
  license:
    name: MIT
    identifier: MIT
  contact:
    name: Infinitude Modern Wrap
    url: https://github.com/vbulkin/infinitude-modern-wrap
servers:
  - url: http://{host}:{port}
    description: Local add-on
    variables:
      host:
        default: local-infinitude
      port:
        default: "3001"
...
paths:
...
  /v1/debug/capture/start:
    post:
      tags: [debug]
      summary: Enable traffic capture.
      description: |
        Flips the in-memory capture flag on. Every subsequent non-excluded
        request (southbound + northbound) has its request body, response
        body, and metadata teed to the `capture_traffic` SQLite table.
        Capture is off on process start; call this endpoint again to
        re-enable after a restart.
      operationId: startCapture
      responses:
        "200":
          description: Current capture status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaptureStatus"
        "503":
          description: Persistence layer unavailable — capture cannot start.
...
  /v1/debug/capture/entries:
    get:
      tags: [debug]
      summary: Paginated metadata listing of captured rows.
      description: |
        Metadata only (no bodies) so responses stay small. Use
        `/v1/debug/capture/entries/{id}` for the full payload of a row.
      operationId: listCaptureEntries
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 1000, default: 100 }
        - in: query
          name: sinceId
          description: Return rows with id strictly greater than this (cursor).
          schema: { type: integer, minimum: 0 }
        - in: query
          name: direction
          schema:
            type: string
            enum: [southbound, northbound, carrier_out]
        - in: query
          name: method
          schema: { type: string }
        - in: query
          name: pathPrefix
          schema: { type: string }
      responses:
        "200":
          description: Entry metadata (newest first).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CaptureEntryMeta"
  /v1/debug/capture/entries/{entry_id}:
    get:
      tags: [debug]
      summary: Single capture row with bodies.
      description: |
        Bodies are returned either as a decoded UTF-8 string (for textual
        content-types) or as base64-encoded bytes (for binary or
        undecodable content). The `reqBodyEncoding` / `respBodyEncoding`
        field tells the consumer which.
      operationId: getCaptureEntry
      parameters:
        - in: path
          name: entry_id
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "200":
          description: Capture entry (metadata + bodies).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaptureEntry"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
...
  responses:
    NotFound:
      description: Resource does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable error code.
              examples: ["validation_error", "not_found", "conflict", "upstream_error"]
            message:
              type: string
            details:
              type: array
              items:
                type: object
                required: [path, issue]
                properties:
                  path:
                    type: string
                    description: JSON pointer into the request body.
                  issue:
                    type: string

    # ── Enums ─────────────────────────────────────────────────────────
...
    IsoDateTime:
      type: string
      format: date-time
      description: |
        ISO 8601 timestamp with an explicit `Z` suffix (UTC).
        The API always emits and accepts `Z`-suffixed UTC timestamps;
        clients render local time for the user.
      pattern: "^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}(?:\\.[0-9]+)?Z$"
      examples: ["2026-04-17T22:30:00Z"]
...
    CaptureStatus:
      type: object
      required:
        [enabled, maxRows, rowCount, oldestAt, newestAt, totalBytes,
         submitted, errors]
      properties:
        enabled: { type: boolean }
        maxRows:
          type: integer
          description: Rolling-window cap — oldest rows trimmed on insert.
        rowCount: { type: integer }
        oldestAt:
          oneOf:
            - $ref: "#/components/schemas/IsoDateTime"
            - type: "null"
        newestAt:
          oneOf:
            - $ref: "#/components/schemas/IsoDateTime"
            - type: "null"
        totalBytes:
          type: integer
          description: Sum of stored request + response body bytes.
        submitted:
          type: integer
          description: Rows attempted since process start (survives flush).
        errors:
          type: integer
          description: Insertion failures since process start.
    CaptureEntryMeta:
      type: object
      required:
        [id, capturedAt, direction, method, path, statusCode,
         reqBytes, respBytes]
      properties:
        id: { type: integer }
        capturedAt: { $ref: "#/components/schemas/IsoDateTime" }
        direction:
          type: string
          enum: [southbound, northbound, carrier_out]
        method: { type: string }
        path: { type: string }
        query:
          type: string
          nullable: true
        statusCode: { type: integer }
        reqContentType:
          type: string
          nullable: true
        reqBytes: { type: integer }
        respContentType:
          type: string
          nullable: true
        respBytes: { type: integer }
        durationMs:
          type: integer
          nullable: true
    CaptureEntry:
      allOf:
        - $ref: "#/components/schemas/CaptureEntryMeta"
        - type: object
          properties:
            reqBody:
              type: string
              nullable: true
              description: |
                Request body. Encoding given by `reqBodyEncoding`: when
                `utf-8`, the string is the decoded body; when `base64`,
                the string is base64-encoded raw bytes.
            reqBodyEncoding:
              type: string
              enum: ["utf-8", "base64"]
              nullable: true
            respBody:
              type: string
              nullable: true
            respBodyEncoding:
              type: string
              enum: ["utf-8", "base64"]
              nullable: true
...